Refactor : refactor modals code
Spliting modals code for better readability and maintenance
